removed duplicate StringUtilTest
